[chore] imgProduct 삭제 전 복구 파일
