+++ .gitignore
@@ -107,7 +107,7 @@ local.properties
 
 ### Intellij+all ###
 # Covers JetBrains IDEs: IntelliJ, RubyMine, PhpStorm, AppCode, PyCharm, CLion, Android Studio, WebStorm and Rider
-# Reference: https://intellij-support.jetbrains.com/hc/en-us/articles/206544839
+# Reference: https://intellij-support.jetbrains.[]com/hc/en-us/articles/206544839
 
 # User-specific stuff
 .idea/**/workspace.xml
