--- src/main/java/com/pofol/main/productTest/list/domain/ProductDto.java
@@ -1,4 +1,0 @@
-package com.pofol.main.productTest.list.domain;
-
-public class ProductDto {
-}
