--- src/main/java/com/pofol/main/testProduct/service/TestProductService.java
@@ -1,10 +1,0 @@
-package com.pofol.main.testProduct.service;
-
-import com.pofol.main.testProduct.domain.TestProductDto;
-
-import java.util.List;
-
-public interface TestProductService {
-  List<TestProductDto> getList() throws Exception;
-  TestProductDto read(Integer pro_id) throws  Exception;
-}
