+++ src/main/resources/mapper/testProductMapper.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8" ?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.pofol.main.testProduct.domain.TestProductDto">
+<mapper namespace="com.pofol.main.imgProduct.domain.TestProductDto">
     <select id="selectAll" resultType="TestProductDto">
         SELECT * FROM protest
     </select>
