+++ src/main/resources/mybatis-config.xml
@@ -5,6 +5,8 @@
 <configuration>
 	<typeAliases>
 		<!-- <typeAlias alias="BoardDto" type="com.fastcampus.ch4.domain.BoardDto"/> -->
-		<typeAlias alias="TestProductDto" type="com.pofol.main.testProduct.domain.TestProductDto"/>
+		<typeAlias alias="TestProductDto" type="com.pofol.main.imgProduct.domain.TestProductDto"/>
+		<typeAlias alias="ProductDto" type="com.pofol.main.product.list.domain.ProductDto"/>
+		<typeAlias alias="EventGroupDto" type="com.pofol.main.product.list.domain.EventGroupDto"/>
 	</typeAliases>
 </configuration>
