+++ src/test/java/com/pofol/main/productTest/testProduct/repository/TestProductRepositoryImplTest.java
@@ -1,13 +1,11 @@
-package com.pofol.main.product.testProduct.repository;
+package com.pofol.main.productTest.testProduct.repository;
 
-import com.pofol.main.testProduct.repository.TestProductRepository;
-import org.assertj.core.api.Assertions;
+import com.pofol.main.imgProduct.repository.TestProductRepository;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
-import org.springframework.test.context.web.WebAppConfiguration;
 
 //@ExtendWith(SpringExtension.class)
 //@ContextConfiguration(locations = { "file:src/main/webapp/WEB-INF/**/root-context.xml", "file:src/main/webapp/WEB-INF/**/servlet-context.xml" })
